requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0
scipy>=1.11.0

# Data augmentation
albumentations>=1.3.0

# Web app
gradio>=4.0.0
plotly>=5.17.0

# Training utilities
tqdm>=4.66.0
tensorboard>=2.14.0

# Image processing
scikit-learn>=1.3.0

# Config
pyyaml>=6.0

# Optional (for Kaggle/Colab)
kaggle>=1.5.16
huggingface-hub>=0.19.0